feat: extract init logic to separate function
