+++ crates/bitwarden-wasm-internal/src/lib.rs
@@ -1,9 +1,11 @@
 mod client;
 mod crypto;
 mod custom_types;
+mod init;
 mod ssh;
 mod vault;
 
 pub use client::BitwardenClient;
 pub use crypto::CryptoClient;
+pub use init::init_sdk;
 pub use vault::{folders::ClientFolders, VaultClient};
